# spec/openapi.yaml
openapi: 3.0.3
info:
  title: ZZAPI MES
  description: SAP ICF REST handlers for MES integration on SAP_BASIS 700
  version: 0.1.0
  license:
    name: UNLICENSED
servers:
  - url: http://sapdev.fastcell.hk:8000
    description: sapdev (Phase 1, direct SAP access)
  - url: http://localhost:8080
    description: zzapi-mes hub (Phase 3)
security: []
paths:
...
  /sap/bc/zzapi_mes_stock:
    get:
      operationId: getStock
      summary: Stock / availability
      description: Returns current stock and ATP availability for a material at a plant.
      parameters:
        - $ref: "#/components/parameters/SapClient"
        - name: matnr
          in: query
          required: true
          schema:
            type: string
            maxLength: 18
          description: Material number
        - name: werks
          in: query
          required: true
          schema:
            type: string
            maxLength: 4
          description: Plant
        - name: lgort
          in: query
          required: false
          schema:
            type: string
            maxLength: 4
          description: Storage location (optional)
      responses:
        "200":
          description: Stock found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockResponse"
        "404":
          description: No stock data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sap/bc/zzapi_mes_wc:
    get:
      operationId: getWorkCenter
      summary: Work center
      description: Returns work center details including capacity and cost center.
      parameters:
        - $ref: "#/components/parameters/SapClient"
        - name: arbpl
          in: query
          required: true
          schema:
            type: string
            maxLength: 8
          description: Work center
        - name: werks
          in: query
          required: true
          schema:
            type: string
            maxLength: 4
          description: Plant
      responses:
        "200":
          description: Work center found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterResponse"
        "404":
          description: Work center not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # -----------------------------------------------------------------------
  # Hub mirrors for Phase 5A endpoints
  # -----------------------------------------------------------------------
...
  /stock/{matnr}:
    get:
      operationId: hubGetStock
      summary: Stock / availability via hub
      description: Proxies to SAP stock handler. Requires `stock` scope.
      security:
        - bearerAuth: []
      parameters:
        - name: matnr
          in: path
          required: true
          schema:
            type: string
            maxLength: 18
          description: Material number
        - name: werks
          in: query
          required: true
          schema:
            type: string
            maxLength: 4
          description: Plant
        - name: lgort
          in: query
          required: false
          schema:
            type: string
            maxLength: 4
          description: Storage location (optional)
      responses:
        "200":
          description: Stock data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient scope (requires 'stock')
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid path parameter (exceeds maxLength)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No stock data found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: SAP upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: SAP request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /work-center/{arbpl}:
    get:
      operationId: hubGetWorkCenter
      summary: Work center via hub
      description: Proxies to SAP work center handler. Requires `work_center` scope.
      security:
        - bearerAuth: []
      parameters:
        - name: arbpl
          in: path
          required: true
          schema:
            type: string
            maxLength: 8
          description: Work center
        - name: werks
          in: query
          required: true
          schema:
            type: string
            maxLength: 4
          description: Plant
      responses:
        "200":
          description: Work center found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkCenterResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient scope (requires 'work_center')
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid path parameter (exceeds maxLength)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Work center not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            retry-after:
              schema:
                type: integer
              description: Seconds until next request is allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: SAP upstream error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: SAP request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # -----------------------------------------------------------------------
  # Phase 5B — MES write-back endpoints
  # -----------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SapClient:
      name: sap-client
      in: query
      required: true
      schema:
        type: integer
        example: 200
      description: SAP client number
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "PO not found"
...
    StockResponse:
      type: object
      required: [matnr, werks]
      properties:
        matnr:
          type: string
          maxLength: 18
          description: Material number
          example: "10000001"
        werks:
          type: string
          maxLength: 4
          description: Plant
          example: "1000"
        items:
          type: array
          description: Stock by storage location/batch
          items:
            $ref: "#/components/schemas/StockItem"
    StockItem:
      type: object
      required: [lgort, clabs]
      properties:
        lgort:
          type: string
          maxLength: 4
          description: Storage location
          example: "0001"
        charg:
          type: string
          maxLength: 10
          description: Batch number
          example: "BATCH001"
        clabs:
          type: number
          description: Unrestricted stock
          example: 250
        avail_qty:
          type: number
          description: Available quantity (ATP)
          example: 200
...
    WorkCenterResponse:
      type: object
      required: [arbpl, werks]
      properties:
        arbpl:
          type: string
          maxLength: 8
          description: Work center
          example: "TURN1"
        werks:
          type: string
          maxLength: 4
          description: Plant
          example: "1000"
        ktext:
          type: string
          description: Work center description
          example: "CNC Turning Center"
        steus:
          type: string
          maxLength: 4
          description: Control key
          example: "PP01"
        kapid:
          type: string
          maxLength: 8
          description: Capacity ID
          example: "10000001"
        kostl:
          type: string
          maxLength: 10
          description: Cost center
          example: "10001000"

    # -----------------------------------------------------------------
    # Phase 5B schemas — write-back
    # -----------------------------------------------------------------
...